+++ sdk/core/azure-core/src/http/curl/curl_session_private.hpp
@@ -182,8 +182,9 @@ namespace Azure { namespace Core { namespace Http {
        * Returning a value smaller than the buffer size will likely indicate that the HTTP
        * RawResponse is completed and that the rest of the buffer contains part of the response
        * body.
+       * 
        */
-      size_t Parse(uint8_t const* const buffer, size_t const bufferSize);
+      int32_t Parse(uint8_t const* const buffer, size_t const bufferSize);
 
       /**
        * @brief Indicates when the parser has completed parsing and building the HTTP RawResponse.
